price 타입 Int -> Long

+++ src/main/kotlin/com/colabear754/jpa_example/domain/entities/item/Album.kt
@@ -2,12 +2,13 @@ package com.colabear754.jpa_example.domain.entities.item
 
 import com.colabear754.jpa_example.dto.member.item.RegistItemRequest
 import com.colabear754.jpa_example.util.typeMismatch
+import jakarta.persistence.Column
 import jakarta.persistence.Entity
 
 @Entity
 class Album(
     name: String,
-    price: Int,
+    price: Long,
     stockQuantity: Int = 1,
     var artist: String,
     var etc: String,
